Fixing test breaks
